+++ syndicate/adapters/base.py
@@ -3,10 +3,7 @@ Interface(s) for adapters subclasses.  Eventually common authentication
 should find its way here.
 """
 
-from __future__ import print_function, division
-
 import collections
-
 
 Response = collections.namedtuple('Response', ('http_code', 'headers',
                                   'content', 'error', 'extra'))
@@ -14,6 +11,14 @@ Response = collections.namedtuple('Response', ('http_code', 'headers',
 
 class AdapterBase(object):
     """ Adapter interface.  Must subclass. """
+
+    def __init__(self, connect_timeout=None, request_timeout=None,
+                 serializer=None, auth=None, ingress_filter=None):
+        self.connect_timeout = connect_timeout
+        self.request_timeout = request_timeout
+        self.serializer = serializer
+        self.auth = auth
+        self.ingress_filter = ingress_filter
 
     def set_header(self, header, value):
         """ Set a header that will be included in every HTTP request. """
@@ -33,3 +38,16 @@ class AdapterBase(object):
 
     def request(self, method, url, data=None, callback=None, query=None):
         raise NotImplementedError('pure virtual method')
+
+
+class AdapterPager(object):
+    """ A sized generator that iterators over API pages. """
+
+    def __init__(self, getter, path, kwargs):
+        self.getter = getter
+        self.path = path
+        self.kwargs = kwargs
+        super().__init__()
+
+    def __len__(self):
+        raise NotImplementedError("pure virtual")